Implement promotion and manager management

+++ Resources/ui/components/tablerow/ManagerRow.js
@@ -2,34 +2,36 @@ var _ = require('lib/underscore');
 
 function ManagerRow (_args) {
     var item = _args.data,
+        deleteHandler = _args.deleteHandler,
+        deleteEnabled = (!item.owner && _args.owner),
         self = Ti.UI.createTableViewRow({
             _id: item.id,
-            height: 80,
+            height: 120,
             className: 'managerInfo'
         });
 
     var imageView = Ti.UI.createImageView({
         left: 0,
-        height: 80,
+        height: 120,
         touchEnabled: false,
-        image: item.photo,
-        width: 80
+        image: item.user.avatar,
+        width: 120
     }),
     nameLabel = Ti.UI.createLabel({
         touchEnabled: false,
         color: '#000',
-        left: 90,
+        left: 130,
         right: 90,
         font: {fontWeight: 'bold', fontSize: 28},
-        text: item.full_name
+        text: item.user.full_name
     }),
     deleteView = Ti.UI.createView({
         right: 10,
         width: 80,
         height: 80,
-        backgroundImage: '/images/trash_blue.png',
-        backgroundFocusedImage: '/images/trash_red.png',
-        backgroundSelectedImage: '/images/trash_red.png'
+        backgroundImage: deleteEnabled ? '/images/trash_blue.png' : '/images/trash_gray.png',
+        backgroundFocusedImage: deleteEnabled ? '/images/trash_red.png' : '/images/trash_gray.png',
+        backgroundSelectedImage: deleteEnabled ? '/images/trash_red.png' : '/images/trash_gray.png'
     });
 
     self.add(imageView);
@@ -37,8 +39,10 @@ function ManagerRow (_args) {
     self.add(deleteView);
 
     deleteView.addEventListener('click', function (e) {
-        e.item_id = item.id;
-        deleteHandler(e);
+        if (deleteEnabled) {
+            e.item_id = item.id;
+            deleteHandler(e);
+        }
     });
 
     return self;
